Fix PR comments
- Handle nagbar visibility inside the component itself
