+++ app/components/Nagbar/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useCallback } from 'react';
+import React, { useState, useMemo, useCallback } from 'react';
 import {
     IoCloseOutline,
 } from 'react-icons/io5';
@@ -6,13 +6,19 @@ import {
     QuickActionButton,
     Link,
 } from '@the-deep/deep-ui';
+import {
+    isNotDefined,
+} from '@togglecorp/fujs';
 
 import useLocalStorage from '#hooks/useLocalStorage';
 
 import styles from './styles.css';
 
 const releaseVersion = process.env.REACT_APP_RELEASE_VERSION;
-const wikiLink = `https://github.com/the-deep/deeper/wiki/${releaseVersion}`;
+const wikiLink = `https://github.com/the-deep/deeper/wiki/DEEP-${releaseVersion}`;
+const today = new Date().getTime();
+const releaseTime = process.env.REACT_APP_RELEASE_TIME;
+const releaseDate = new Date(releaseTime ?? '').getTime();
 
 function Nagbar() {
     const [
@@ -21,18 +27,32 @@ function Nagbar() {
     ] = useState<boolean>(true);
 
     const [
-        releaseNoteSeen,
-        setReleaseNoteSeen,
-    ] = useLocalStorage<boolean>('release-note-seen', false);
+        seenReleaseNote,
+        setSeenReleaseNote,
+    ] = useLocalStorage<string | undefined>('release-note-seen', undefined);
+
+    const showNagbar = useMemo(() => {
+        if (isNotDefined(releaseTime) || !nagbarShown) {
+            return false;
+        }
+        const diff = Math.round((today - releaseDate) / (60 * 60 * 24 * 1000));
+        if (diff < 0 || diff > 15) {
+            return false;
+        }
+
+        return true;
+    }, [
+        nagbarShown,
+    ]);
 
     const handleClose = useCallback(() => {
         setNagbarShown(false);
-        setReleaseNoteSeen(true);
+        setSeenReleaseNote(releaseVersion);
     }, [
-        setReleaseNoteSeen,
+        setSeenReleaseNote,
     ]);
 
-    if (!nagbarShown || releaseNoteSeen) {
+    if ((seenReleaseNote === releaseVersion) || !showNagbar) {
         return null;
     }
 
@@ -40,17 +60,16 @@ function Nagbar() {
         <div className={styles.nagbar}>
             <div className={styles.content}>
                 DEEP has been updated to
-                &nbsp;
+                &#8204;
                 {releaseVersion}
                 .
-                &nbsp;
-                Read
-                &nbsp;
+                &#8204;
                 <Link
                     to={wikiLink}
                     actionsContainerClassName={styles.linkIcon}
+                    className={styles.link}
                 >
-                    release notes
+                    Find the release notes here
                 </Link>
                 .
             </div>
@@ -59,6 +78,7 @@ function Nagbar() {
                 name={undefined}
                 onClick={handleClose}
                 variant="transparent"
+                title="Dismiss"
             >
                 <IoCloseOutline />
             </QuickActionButton>
