+++ app/components/Nagbar/styles.css
@@ -2,16 +2,21 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    background-color: var(--dui-color-complement3);
+    background-color: var(--dui-color-complement2);
     width: 100%;
-    gap: var(--dui-spacing-extra-small);
+    color: var(--dui-color-brand);
 
     .content {
         display: flex;
         padding: var(--dui-spacing-extra-small);
     }
 
-    .link-icon {
-        display: none;
+    .link {
+        text-decoration: underline;
+        color: var(--dui-color-brand);
+
+        .link-icon {
+            display: none;
+        }
     }
 }
